# Conflicts:
#	app/src/main/java/com/oreo/ui/custom/HRCombinedChart.kt

+++ app/src/main/java/com/oreo/data/model/HRModel.kt
@@ -5,5 +5,5 @@ data class HRModel(
     val maxValues: Int,
     val minValues: Int,
     val values: List<Int>? = null,
-    val midValues: Float,
+    val midValues: Int,
 )
